Check if files exist before removing them

+++ Jenkinsfile
@@ -13,12 +13,15 @@ node(NODE_NAME)
         {
             checkout scm
         }
-        stage("Remove existing dist and docs")
+        if (Files.exists("_dist") || Files.exists("_docs"))
         {
-            bat """
-                rmdir /Q /S "_dist"
-                rmdir /Q /S "_docs"
-            """
+            stage("Remove existing dist and docs")
+            {
+                bat """
+                    rmdir /Q /S "_dist"
+                    rmdir /Q /S "_docs"
+                """
+            }
         }
         for (int i = 0; i < 5; i++) {
             stage("Stage ${i}") {
